moving function within package
