+++ db/start_database.go
@@ -2,6 +2,7 @@ package db
 
 import (
 	"fmt"
+	"log"
 	"net"
 	"os"
 	"os/exec"
@@ -71,6 +72,19 @@ func (slt Invoker) IsValid() error {
 		return nil
 	}
 	return fmt.Errorf("Invalid invoker. Can be one of '%v', '%v', '%v', '%v' or '%v'", InvokerService, InvokerQuery, InvokerInstaller, InvokerPlugin, InvokerCheck)
+}
+
+// StartImplicitService starts up the service in an implicit mode
+func StartImplicitService(invoker Invoker, refreshConnections bool) error {
+	utils.LogTime("db.StartImplicitService start")
+	defer utils.LogTime("db.StartImplicitService end")
+
+	log.Println("[TRACE] start implicit service")
+
+	if _, err := StartDB(constants.DatabaseDefaultPort, ListenTypeLocal, invoker, refreshConnections); err != nil {
+		return err
+	}
+	return nil
 }
 
 // StartDB :: start the database is not already running
